Fix compiler warnings appearing with VS2008. Based on Steven Van Ingelgem's patch in #2787839 with a bunch of additional edits by yours truly.

+++ src/packet.c
@@ -1078,7 +1078,7 @@ _libssh2_packet_require(LIBSSH2_SESSION * session, unsigned char packet_type,
             return ret;
         } else if (ret == 0) {
             /* nothing available, wait until data arrives or we time out */
-            long left = LIBSSH2_READ_TIMEOUT - (time(NULL) - state->start);
+            long left = LIBSSH2_READ_TIMEOUT - (long)(time(NULL) - state->start);
 
             if (left <= 0) {
                 state->start = 0;
@@ -1187,7 +1187,7 @@ _libssh2_packet_requirev(LIBSSH2_SESSION * session,
             return ret;
         }
         if (ret <= 0) {
-            long left = LIBSSH2_READ_TIMEOUT - (time(NULL) - state->start);
+            long left = LIBSSH2_READ_TIMEOUT - (long)(time(NULL) - state->start);
 
             if (left <= 0) {
                 state->start = 0;
